pyproject.toml
[tool.uv]
package = false
#===========================
[project]
name = "ocpy10"
version = "0.1.0"
description = "API RESTful SoftDesk"
authors = [
    { name = "Arnaud Goguelin", email = "arnaud.goguelin@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "django ~= 5.2.8",
    "djangorestframework ~= 3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "djangorestframework-stubs>=3.16.6",
    "drf-spectacular ~= 0.29.0",
    "pygments ~= 2.19.2",
    "python-dotenv>=1.2.1",
    "pyyaml ~= 6.0.3",
    "uritemplate ~= 4.2.0",
    "uvicorn ~= 0.38.0",
]
readme = "README.md"
#===========================
[dependency-groups]
dev = [
    "faker>=40.1.2",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
    "pytest-mock>=3.15.1",
    "ruff ~= 0.12.4",
]
#===========================
[tool.ruff]
line-length = 120
target-version = "py312"
include = ["*.py", "*.pyi"]
exclude = [
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".tox",
    ".venv",
    ".envTemplate",
    "_build",
    "buck-out",
    "build",
    "dist",
    "compose.yml",
    "Dockerfile",
    ".gitignore",
    "__pycache__",
    "old",
    "__init__.py",
    "*.pyc",
    "*/migrations/",
    "*/settings/*"
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "Q",      # flake8-quotes
]

ignore = []

[tool.ruff.format]
line-ending = "auto"
indent-style = "space"
quote-style = "double"
skip-magic-trailing-comma = false
docstring-code-format = true
docstring-code-line-length = 120

[tool.ruff.lint.flake8-quotes]
inline-quotes = "double"
multiline-quotes = "double"
docstring-quotes = "double"

[tool.ruff.lint.isort]
known-first-party = ["src"]
force-single-line = false
lines-between-types = 1
lines-after-imports = 2

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.local"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
